Use OiNutter’s actively maintained nodenv package
Use this one instead of wfarr’s, which hasn’t been touched in over a year. This change also fixes a problem where the `npm` binary wasn’t visible immediately after selecting a global node version – it required a new shell to be opened. This problem doesn’t exist in OiNutter’s version of nodenv.

This is the last fix needed for a 100% smooth and error-free first run of babushka.